Move exceptions up a level
