+++ app.py
@@ -3,7 +3,7 @@ import logging
 import streamlit as st
 
 from sfta.context_manager import DeepRecurse
-from sfta.data_models.exceptions.base import FaultTreeTextException
+from sfta.exceptions.base import FaultTreeTextException
 from sfta.core.fault_tree import FaultTree
 from sfta.output_handling import create_directory_robust, write_output_files
