+++ sfta/core/gate.py
@@ -2,7 +2,7 @@ import re
 from .utilities import descending_product, descending_sum, Nan, is_bad_id
 from .tome import Tome, Writ
 from .event import Event
-from sfta.data_models.exceptions.gate import (
+from sfta.exceptions.gate import (
     GateBadIsPagedException,
     GateBadTypeException,
     GateCommentAlreadySetException,
@@ -16,7 +16,7 @@ from sfta.data_models.exceptions.gate import (
     GateTypeNotSetException,
     GateZeroInputsException,
 )
-from sfta.data_models.exceptions.fault_tree import FaultTreeBadIdException
+from sfta.exceptions.fault_tree import FaultTreeBadIdException
 
 
 class Gate:
