+++ sfta/exceptions/base.py
