+++ sfta/exceptions/event.py
