+++ sfta/exceptions/fault_tree.py
