+++ sfta/exceptions/gate.py
